requirements.txt
# Run the following
pip install benepar
pip install spacy
python import benepar
benepar.download('benepar_en3')
exit()
